+++ src/classes/parse/cdd_cst_type_eval.c
@@ -143,7 +143,7 @@ static cdd_c_error_t extract_type_name(cdd_cst_node_t *node, char **out_name,
 
   if (buf_len > 0) {
     char *ret;
-    ret = (char *)C_CDD_MALLOC(buf_len + 1);
+    ret = (char *)(size_t)C_CDD_MALLOC(buf_len + 1);
     if (!ret) {
       C_CDD_LOG_DEBUG("ENOMEM: OOM\n");
       return CDD_C_ERROR_MEMORY;
